Moved downloads handling to separate (crashes!)
The downloading of data have been moved to a separate DownloadManager.
This manager replaces the old QNetworkAccessManager. The Installer class
interfaces the DownloadManager through the signals
* downloadComplete()
* partialData()
and initiates a download with
* get()

It works for getting SF XML data, but has trouble downloading the actual
image. Some Sourceforge mirrors transfers all data in parts (using the
raw http header "range") while others transfers it through one big http
response. The latter way does work, but the installer crashes when done.

This, and other cleanups and fixes, needs to be done before the next
release!