Moved image writing to a separate worker thread
This allows the UI to still be responsive even though the diskwriter is
doing heavy IO. Especially on some versions of GNU/Linux where all write
operations seem to be cached and only written to the SD card when the
file handle is closed. This means that no GUI updates were processed
until all data has been written to the SD card. Now, since all image
writing is done in a separate thread, the GUI can still be responsive.
Basically, it just looks better now.